[SYCL][NFC] Move SizeListToStr to utilities
This commit moves the utility from the kernel properties headers to the
common property utiliites headers. This will allow for use in other
properties that seek to use multiple integral values in their
compile-time values.
